Damage and Dice
The new damage die side is functional. Damage die rolls add to the
damage stat. The damage stat is added to an attack's damage and then
depleted. Dice now add value to a stat when rolling is completed. Die
rolling is complete when the roll button is pressed with 6 dice or after
the 3 roll.
attacks that deal damage now use the applyDamage() method instead of
calcDamage(). applyDamage is calcDamage() but also depletes a monster's
damage stat.