Die set introduced, New Status effect type possible
DieSet is a new class which represents a set of dice. Die set will in
the future be helpful in describing required dice for monster attacks.
Currently it is used as a savingRoll for Virgeo's attack Bramblesnare's
new function. Bramble snare now inflicts a status effect--ensnared.
Ensnared deals 5 Combat damage a turn. This can be removed by the
inflicted player by rolling 3 combat dice. Damage or effect removal
happens at the end of the players turn.